requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
beautifulsoup4==4.13.5
dbt-core==1.10.11
dbt-postgres==1.9.1
pandas==2.1.4
pandera[io]==0.26.1
# psycopg2-binary==2.9.10
# pydantic>=2.11.7
# pytest>=8.4.1
python-dotenv==1.1.1
rapidfuzz>=3.13.0
requests==2.32.5
# sqlalchemy>=2.0.42
unidecode==1.4.0
astronomer-cosmos==1.10.1
selenium
apache-airflow-providers-amazon
